refactor(codec): update LZ4 & memory internals
